Friction Generator
This Python script was conceived while reading Atomic Habits by James Clear.

It is an application of the inversion of his 3rd law of habit change, "make it easy". The inversion of the law is "make it difficult". Specifically, to create as much tension as possible to bad habits- making it hard to do them. Imagine putting your TV in your garage after every use, or your remote control in your car. We can reduce the temptation, or rather ability, to execute certain habits by adding as much "friction" as possible to doing them.

This script is an application of this principle. Netflix has been used as the token bad habit in this instance. The script blocks the user entry to Netflix's website, unless a correct password is given. This password however, is a randomly generated 6-letter word which corresponds to 6 numeric values, which must be solved for through 6 algebraic problems.  The values are then ran through a final formula which correspond to the randomly generated 6-letter word with 5 operators.

The user must get every answer right every time or else they are prompted to start from scratch. The entire system is designed to be a hassle, thus fulfilling James' inverted 3rd law properly, and ensuring the user will not use Netflix at times when they really shouldn't be- doing the task being procrastinated on is likely to be less of a hassle than deciphering the encryption.

The script is a work in progress and more points of friction can be added, such as a short timer for each problem, along with bug fixes.